.data: support fine-grained filtering of activity types
The code as it currently stands extracts data for all activity types instead
of just a subset. This change makes it possible to filter by activity
type, reducing the amount of data that needs to be pulled down from the
server.

This is helpful if producing GPX files--which only apply to some
activities.

Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>